164797982-feature(comment): rename article reading card
